Move calculation of merger prob rk into as_merge()
To enable the merger probability to be stored for each node, the
merger log probability log_rk is now calculated within as_merge
and provided as an attribute when creating the new, merged, node.
Nodes created not by merger (i.e. leaves) have None for log_rk.
This creates a small problem in that it is no longer so easy to
obtain the log marginal likelihood for the merger tree.