Correct calculation of log marginal likelihood for trees
Previously the definition of the log marginal likelihood being used
in the code was wrong as it was not recurssing down to include all
possible subtrees as in eqn 2 of Heller & Ghahramani (2005).
I have now corrected this by having as_merge() calculate the
correct log_ml by using the log_ml of each child node.

Note that this update makes the log_marginal_likelihood() method
unnecessary, so it has been removed.