Add dist functions for observations with noise
This adds a class uncert_NormalFixedCovar that provides probability
functions for data with Gaussian errors, subject to a Gaussian prior
on cluster mean, with cluster covariances assumed known.
A NIW prior on noisy data can not be used since the parameters
of the resulting posterior etc cannot be found analytically.