Correct two bugs
The first bug corrected was that for very large alpha the
calculation of log_pi in refine_probs() was causing OverflowErrors.
This was corrected with a (safe!) numerical approximation.

Also, when merging the GMM from true_bhc s I was not updating the
number of components in the GMM. This would mean that gmm_mean and
gmm_covar would not be calculated properly. This too has been
corrected.