Replace get_individual_potseriors()
The method get_individual_posteriors() involved storing a gmm.GMM
for each datum. This would quickly take up a large amount of
memory.

Now the posterior for a given datum is instead calculated on request
and returned, but not stored. This takes the same amount of CPU,
but saves memory