Add target_prob param to tree_posterior_predictive_prob()
This adds the option of noting what the probability of a comparison
case is when running tree_posterior_predictive_prob(). When this
probability is exceeded we can stop the calculation in
tree_posterior_predictive_prob() as we know the probability it
finds will be larger that what it is comparing to.

This is used in noisy_rbhc_node.filter_data() to considerably
shorten computation time