Refactor noisy_bhc init-ing
This moves the creation and merger of nodes into a tree out of
noisy_bhc.__init__() and into seperate functions. A factory function
noisy_bhc.from_data() provides the previous functionality of
__init__() and has been inserted into noisy_rbhc.py where needed.

This has been done so that the structure is in place to allow
bhc trees to be built from structures other than nodes containing
single data - specifically, from some prior clustering.