Correct calculation of nk & so log_pi
WHen clustering clusters - the value of nk, which is used in the
calculation of the log-prior probability of joining two clusters
(log_pi), should be the number of preclusters within a cluster,
not the number of data points.

By setting this correctly we get sensible values of log_pi and so
reasonable clusterings.